Rename consent.ts -> initOverlay.ts
