+++ clients/fides-js/src/fides.ts
@@ -193,7 +193,7 @@ if (typeof window !== "undefined") {
 // Export everything from ./lib/* to use when importing fides.mjs as a module
 export * from "./services/api";
 export * from "./services/external/geolocation";
-export * from "./lib/consent";
+export * from "./lib/initOverlay";
 export * from "./lib/consent-context";
 export * from "./lib/consent-types";
 export * from "./lib/consent-utils";
